Move security tests from other to security
